Remove main.h and create Common.cpp and Common.h moved to Common:  * String mqttServer  * String mqttPort  * String mqttUserName  * String mqttPassword  * bool rebootFlag  * readRebootFlag()  * writeRebootFlag()  * readConfigFile()  * writeConfigFile()
