--- src/main.h
@@ -1,6 +1,0 @@
-#ifndef MAIN_H
-#define MAIN_H
-
-void writeRebootFlag(bool flagValue);
-
-#endif
